add useEndpoint hook
